Move PartitionProjection utilities into PartitionProjectionProperties
